Feature:
the main is separated to the main.rb file
