+++ main.rb
@@ -1,143 +1,29 @@
 require_relative 'app'
 
-class Nameable
-  def correct_name
-    raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
+def start_method(app, user_option)
+  methods_list = {
+    1 => :list_all_books,
+    2 => :list_all_people,
+    3 => :create_person,
+    4 => :create_book,
+    5 => :create_rental,
+    6 => :list_person_rental,
+    7 => :exit
+  }
+  app.send(methods_list[user_option])
+end
+
+def main
+  app = App.new
+  puts 'Welcome to School Library App!'
+  loop do
+    option = present_options
+    if option == 7
+      puts 'Thank you for using this app!'
+      break
+    end
+    start_method(app, option)
   end
 end
 
-class Person < Nameable
-  attr_reader :id
-  attr_accessor :name, :age, :rentals
-
-  def initialize(age, name = 'Unknown', parent_permission: true)
-    super()
-    @id = rand(1..1000)
-    @name = name
-    @age = age
-    @parent_permission = parent_permission
-    @rentals = []
-  end
-
-  def can_use_services?
-    of_age? || @parent_permission
-  end
-
-  def correct_name
-    @name
-  end
-
-  def add_rentals(book, date)
-    @rentals << Rental.new(date, book, self)
-  end
-
-  private
-
-  def of_age?
-    @age >= 18
-  end
-end
-
-class BaseDecorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    original_name = @nameable.correct_name
-    original_name.capitalize
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name[0..9]
-  end
-end
-
-person = Person.new(age: 22, name: 'maximilianus', parent_permission: true)
-person.correct_name
-capitalized_person = CapitalizeDecorator.new(person)
-puts capitalized_person.correct_name
-capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
-puts capitalized_trimmed_person.correct_name
-
-class Student < Person
-  attr_reader :classroom
-
-  def initialize(age, name, classroom, parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    '¯(ツ)/¯'
-  end
-
-  def classroom=(classroom)
-    @classroom = classroom
-    classroom.students.push(self) unless classroom.students.include?(self)
-  end
-end
-
-class Classroom
-  attr_accessor :label
-  attr_reader :students
-
-  def initialize(label)
-    @label = label
-    @students = []
-  end
-
-  def add_student(student)
-    @students << student
-    student.classroom = self
-  end
-end
-
-class Teacher < Person
-  def initialize(age, name, specialization, parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
-    @specialization = specialization
-  end
-
-  def can_use_services?
-    true
-  end
-end
-
-class Book
-  attr_accessor :title, :author, :rentals
-
-  def initialize(title, author)
-    @title = title
-    @author = author
-    @rentals = []
-  end
-
-  def add_rental(person, date)
-    @rentals << Rental.new(date, self, person)
-  end
-end
-
-class Rental
-  attr_accessor :date, :book, :person
-
-  def initialize(date, book, person)
-    @date = date
-    @book = book
-    @person = person
-
-    book.rentals << self
-    person.rentals << self
-  end
-end
+main
